requirements.txt
python-gitlab
requests